Add k=2 option for fixedsplines_tp
Problem: using lc_linear worked, but DIC was on a completeky different scale compared to the splines DIC, so we couldn't possibly compare DIC values.
Strategy: In order to keep DIC on the same scale, we keep the "jagam" approach used in fixedsplines_tp for k = 3 and upwards, but constrain it to making only linear fits. Using jagam for k >= 3, jagam outputs an "X" array with k columns, where the first column is a constant, the last column is the linear component, and the columns between (for k = 3, only column 2) is a spline component. The procedure fits "b" coefficients to each of these components, so for k = 3, b[1] corresponds to intercept, b[3] corresponds to the slope, and b[2] is the non-linear component. So to get a linear model (b = 2), we make a b = 3 JAGAM object, but manipulate the jagam object and the JAGS code so as to constrain b[2] = 0, leaving only two b's, b[1] and b[3], to be "fitted" by JAGS (these b's are called 'b_new' in the JAGS code).

Changes to leftcensored_fixedsplines_tp.R file:

Changes to function lc_fixedsplines_tp:
    - Line 41: Add 'k_jagam' as internal parameter. When we want a linear model, k_jagam is
    - Line 92: We use both k_jagam and k and in the call to get_jagam_object. (see below).
    - Line 102: We still use the original k in the call to get_jags_model_code (but the name of the argument has been chnaged to b_code to avoid/reduce confusion)
    - Line 36 + 157: unelated to the 'b = 2' stuff

Changes to function get_jagam_object:
    - Line 217: The old 'k' parameter is replaced by k_jagam, and we add 'k_orig' which is the k used by the user when calling 'lc_fixedsplines_tp' (for a linear model: k = 2, k_jagam = 3)
    - Line 222: 'k_jagam' is used when we make the gam formula, so indirectly in the call to jagam(). I.e., jagam() makes the output for k = 3 even when we want k = 2, so we later manipulate the jagam object and use a different model. (Calling jagam with k = 2 or without a spline effect results in an error, that's why we have to call it with k = 3.)
    - Line 244: we add a new parameter 'b2' to the data and set it to zero. This will be used as the b_new[2] parameter in the model.
    - Line 245: Initial values are set for 3 b's, we reduce this to two: number 1 (intercept) and number 3 (slope)

Changes to jags_model_code.R:
Changes to get_jags_model_code:
    - Line 3: change k to k_code to be more specific
    - Line 10-37: add k = 2 model. This is identical to the k = 3 code but has 3 changes: (1) We change it to have only 2 b's instead of 3, changing indices in S1 (line 27), and changing the index 'b[2:3]' to 'b[2] (line 28). (2) We introduce 'b_new', which is the b that actually is going to be used when calculating mu. 'b_new' has fitted values on places 1 and 3 (intercept and slope), while on place 2 we use the given 'b2' value, see 'line 244' above, which is set to zero. I.e., the second column of the X object is ignored.